Split of redundent plotting fuctions
These functions will be removed in the near future.
